Move ModelRef to the model types file.
