+++ ui/src/app/store/types/model.ts
@@ -1,3 +1,10 @@
 export type Model = {
   id: number;
 };
+
+/**
+ * A named foreign model reference, e.g. machine.domain
+ */
+export type ModelRef = Model & {
+  name: string;
+};
